turned tabs component to Routes and reorganized Routes/app.tsx
