+++ App.tsx
@@ -10,26 +10,30 @@ import {
 // Do I need to import platform for react-native?? seems to be working fine w/o it.
 import { useSafeAreaInsets, SafeAreaProvider} from 'react-native-safe-area-context';
 import { NavigationContainer } from '@react-navigation/native'
-import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import SplashScreen from 'react-native-splash-screen'
-import { Tabs } from './Tabs'
-import Register from './pages/register/Register'
-import Login from './pages/login/Login'
-import { ForgotPassword } from './pages/forgotPassword/ForgotPassword';
+import { Routes } from './src/Routes'
 import EStyleSheet from 'react-native-extended-stylesheet'
+import store from './src/redux/store'
+import { Provider } from 'react-redux'
+
 
 function App(): JSX.Element {
+  const theme = {
+    // ...DefaultTheme,
+    // colors: {
+    //   ...DefaultTheme.colors,
+    //   background: '#fff',
+    // },
+  }
   // const isDarkMode = useColorScheme() === 'dark';
-  const Stack = createNativeStackNavigator()
-  
+
   useEffect(()=> {
-    // do stuff while splash screen is shown After having done stuff (such as async tasks) hide the splash screen
     setTimeout(() => {
       SplashScreen.hide();
     }, 3000);
   },[])
   
-  const CustomStatusBar = ({backgroundColor}) => { 
+  const CustomStatusBar = (backgroundColor: any) => { 
     const insets = useSafeAreaInsets();
     return (
       <View style={{ height: insets.top, backgroundColor }}>
@@ -41,18 +45,20 @@ function App(): JSX.Element {
     );
   }
 
+
   return (  
-    <SafeAreaProvider>
-      <NavigationContainer>
-        <CustomStatusBar backgroundColor={EStyleSheet.value('$mainColor_magenta')} />
-        <Stack.Navigator screenOptions={{headerShown: false}} >
-          <Stack.Screen name="Login" component={Login} />
-          <Stack.Screen name="Register" component={Register} />
-          <Stack.Screen name="Home" component={Tabs} />
-          <Stack.Screen name="Forgot_Password" component={ForgotPassword} />
-        </Stack.Navigator>
-      </NavigationContainer>
-    </SafeAreaProvider>
+    // <SafeAreaProvider>
+      <Provider store={store}>
+        <NavigationContainer 
+          // theme={theme}      
+        >
+          <CustomStatusBar backgroundColor={EStyleSheet.value('$mainColor_magenta')} />
+          <Routes 
+            // navigation={undefined} state={undefined} descriptions={undefined} 
+          />
+        </NavigationContainer>
+      </Provider>
+  //  </SafeAreaProvider>
   );
 }
