+++ package.json
@@ -16,6 +16,7 @@
     "@react-navigation/bottom-tabs": "^6.5.8",
     "@react-navigation/native": "^6.1.7",
     "@react-navigation/native-stack": "^6.9.13",
+    "@reduxjs/toolkit": "^1.9.7",
     "axios": "^1.4.0",
     "firebase": "^10.1.0",
     "formik": "^2.4.2",
@@ -31,6 +32,7 @@
     "react-native-screens": "^3.22.0",
     "react-native-splash-screen": "^3.3.0",
     "react-native-vector-icons": "^9.2.0",
+    "react-redux": "^8.1.3",
     "styled-components": "^5.3.1",
     "yup": "^1.2.0"
   },
