+++ src/redux/store.js
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit'
+import userSlice from './userSlice'
+
+export default configureStore({
+  reducer: {
+    user: userSlice,
+  },
+  middleware: getDefaultMiddleware => getDefaultMiddleware({serializableCheck: false}),
+})
