+++ src/screens/Home.tsx
@@ -12,7 +12,7 @@ import {
 import Geolocation from 'react-native-geolocation-service'
 import { useState, useEffect } from 'react'
 import axios from 'axios'
-import { locationPermission } from '../utils/permissions';
+import { locationPermission } from '../../utils/permissions';
 import List from '../components/list/list'
 import Icon from 'react-native-vector-icons/FontAwesome';
 import auth from '@react-native-firebase/auth'
