+++ src/screens/register/registerStyles.js
